#359: create component HeaderPage
