+++ webapp/app/containers/DashboardByUser/HeaderPage/index.js
@@ -1,0 +1,21 @@
+//
+// Container header dashboard by user page
+//
+
+import { connect } from 'react-redux';
+import { DashboardByUserHeaderPage } from 'components/DashboardByUser/HeaderPage';
+
+function mapStateToProps() {
+  return {
+  };
+}
+
+function mapDispatchToProps() {
+  return {
+  };
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(DashboardByUserHeaderPage);
